lambda/requirements.txt
pyjwt==2.0.1